Extract Prompt Studio constants and utils
